+++ app-compose/src/main/kotlin/com/pitchedapps/frost/components/FrostComponents.kt
@@ -14,11 +14,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.pitchedapps.frost.hilt
+package com.pitchedapps.frost.components
 
-import com.pitchedapps.frost.components.Core
-import com.pitchedapps.frost.components.FrostDataStore
-import com.pitchedapps.frost.components.UseCases
 import javax.inject.Inject
 import javax.inject.Singleton
