+++ app-compose/src/main/kotlin/com/pitchedapps/frost/main/MainScreenViewModel.kt
@@ -21,10 +21,10 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
+import com.pitchedapps.frost.components.FrostComponents
 import com.pitchedapps.frost.ext.GeckoContextId
 import com.pitchedapps.frost.ext.idData
 import com.pitchedapps.frost.ext.toContextId
-import com.pitchedapps.frost.hilt.FrostComponents
 import com.pitchedapps.frost.web.state.FrostWebStore
 import com.pitchedapps.frost.webview.FrostWebComposer
 import dagger.hilt.android.lifecycle.HiltViewModel
