+++ app-compose/src/main/kotlin/com/pitchedapps/frost/web/usecases/HomeTabsUseCases.kt
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.pitchedapps.frost.components.usecases
+package com.pitchedapps.frost.web.usecases
 
 import com.pitchedapps.frost.ext.WebTargetId
 import com.pitchedapps.frost.facebook.FbItem
